+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [4.2.4] - 10.03.2025
 
 ### Added